+++ include/limine_manager/config/config.hpp
@@ -30,6 +30,8 @@ struct AppConfig {
     std::vector<std::string> exclude_kernels;
     std::vector<std::string> kernel_order;
 
+    std::string theme_name{"none"};
+
     std::map<std::string, std::string> limine_options{{"remember_last_entry", "yes"},
                                                       {"timeout", "5"}};
 };